+++ border-async-trainer/src/async_trainer/config.rs
@@ -9,23 +9,29 @@ use std::{
 /// Configuration of [AsyncTrainer](crate::AsyncTrainer)
 #[derive(Clone, Debug, Deserialize, Serialize)]
 pub struct AsyncTrainerConfig {
+    /// The maximum number of optimization steps.
+    pub max_opts: usize,
+
     /// Where to save the trained model.
     pub model_dir: Option<String>,
-
-    /// Interval of recording in training steps.
-    pub record_interval: usize,
 
     /// Interval of evaluation in training steps.
     pub eval_interval: usize,
 
-    /// The maximal number of training steps.
-    pub max_train_steps: usize,
+    /// Interval of flushing records in optimization steps.
+    pub flush_record_interval: usize,
+
+    /// Interval of recording agent information in optimization steps.
+    pub record_compute_cost_interval: usize,
 
     /// Interval of saving the model in optimization steps.
     pub save_interval: usize,
 
     /// Interval of synchronizing model parameters in training steps.
     pub sync_interval: usize,
+
+    /// Warmup period, for filling replay buffer, in environment steps
+    pub warmup_period: usize,
 }
 
 impl AsyncTrainerConfig {
@@ -42,5 +48,11 @@ impl AsyncTrainerConfig {
         let mut file = File::create(path)?;
         file.write_all(serde_yaml::to_string(&self)?.as_bytes())?;
         Ok(())
+    }
+
+    /// Sets the directory the trained model being saved.
+    pub fn model_dir<T: Into<String>>(mut self, model_dir: T) -> Result<Self> {
+        self.model_dir = Some(model_dir.into());
+        Ok(self)
     }
 }