+++ border-core/src/replay_buffer/step_proc.rs
@@ -48,15 +48,17 @@ where
         let batch = if self.prev_obs.is_none() {
             panic!("prev_obs is not set. Forgot to call reset()?");
         } else {
+            let is_done = step.is_done();
             let next_obs = step.obs.clone().into();
             let obs = self.prev_obs.replace(step.obs.into()).unwrap();
             let act = step.act.into();
             let reward = step.reward;
-            let is_done = step.is_done;
+            let is_terminated = step.is_terminated;
+            let is_truncated = step.is_truncated;
             let ix_sample = None;
             let weight = None;
 
-            if is_done[0] == 1 {
+            if is_done {
                 self.prev_obs.replace(step.init_obs.into());
             }
 
@@ -65,7 +67,8 @@ where
                 act,
                 next_obs,
                 reward,
-                is_done,
+                is_terminated,
+                is_truncated,
                 ix_sample,
                 weight,
             }