+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 * Fix a bug in `MlpConfig` (`border-tch-agent`)
 * Bump the version of tch to 0.10.0 (`border-tch-agent`)
 * Change the name of trait `StepProcessorBase` to `StepProcessor` (`border-core`)
+* Change the environment API to include terminate/truncate flags (`border-core`) (https://github.com/taku-y/border/issues/10)
 
 ## v0.0.6 (2023-09-19)